Apply Dependency Inversion Principle [LO-DIP] Make Logic and StorageFile depend on Storage interface.
